move accordion and features to index folder
