+++ site/src/pages/_index/accordion/PageIllustration.tsx
