+++ site/src/pages/index.tsx
@@ -10,9 +10,9 @@ import ArrowsIcon from "@site/static/img/arrows.svg";
 import { Button } from "@jpmorganchase/uitk-core";
 
 import styles from "./index.module.css";
-import Features from "../components/features/Features";
+import Features from "./_index/features/Features";
 import Card, { CardProps } from "../components/card/Card";
-import Accordion from "../components/accordion/Accordion";
+import Accordion from "./_index/accordion/Accordion";
 
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
